requirements.txt
asn1crypto==1.3.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
enum34==1.1.10
Flask==1.1.2
Flask-DebugToolbar==0.11.0
Flask-Ext==0.1
Flask-Markdown==0.3
flask-msearch==0.2.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
gunicorn==20.0.4
idna==2.10
ipaddress==1.0.23
itsdangerous==1.1.0
jeepney==0.4.3
Jinja2==2.11.2
keyring==21.2.1
keyrings.alt==3.4.0
Markdown==3.2.2
MarkupSafe==1.1.1
psycopg2==2.8.5
pycparser==2.20
pycrypto==2.6.1
python-dotenv==0.13.0
requests==2.24.0
SecretStorage==3.1.2
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.25.9
Werkzeug==1.0.1
Whoosh==2.7.4
